move Response.js logic to Application.js
