+++ app/lib/my-http-wrapper/http/index.js
@@ -1,2 +1,1 @@
 exports.Application = require('./Application').Application;
-exports.Response = require('./Response').Response;
